requirements.txt
boto3
snowflake-connector-python
python-dotenv